// src/proto/generated/book/Book.ts
// Original file: src/proto/book.proto


export interface Book {
  'id'?: (number);
  'title'?: (string);
  'author'?: (string);
  'isbn'?: (string);
  'publishYear'?: (number);
}

export interface Book__Output {
  'id': (number);
  'title': (string);
  'author': (string);
  'isbn': (string);
  'publishYear': (number);
}